Change definition of excess h2.
Excess h2 in the now refers to the multiples of the required stoichiometric
h2 from fe2o3 to fe, even if only partial reduction occurs in a device.
Using the old definition, we could no garantee that the required gas
oxidation degree is sufficient to reach high metallisation for the DRI-Plasma
device.